retrieve user cookies in user context, not layout
